Start the supervisor daemon
Under a clean Ubuntu 16.04 environment, supervisord isn't started by the
time we attempt to reload its configuration. This results in a somewhat
cryptic error:

   error: <class 'socket.error'>, [Errno 2] No such file or directory: file: /usr/lib/python2.7/socket.py line: 228

It doesn't seem to be used before that point, so we start it then and
that also removes the need to reload the configuration.

Unverified: The assumption that it would start automatically was
probably broken sometime between 3.0r1-1 (version packaged in wily) and
3.2.0 (version packaged in xenial).