Relabel volumes
Hitting permission errors otherwise, could be specific to either the
host parent directory of the bindmount, the one from the container tree
or even special SE overrides configured in /etc -- not sure which. In
any case, we relabel systematically on container creation here.

This is OK for PG but actually not for Orthanc Storage as the sheer
amount of files it can contain renders the operation extremely costly
depending on the size of the server (sometimes trigger Docker API client
timeouts when creating the container).

Note: Even using :z can be too slow as the Docker engine still has to
*check* that the labels are correct. Since :Z is thus not more expensive
we just use it until we can find a proper solution.