Don't specify version explicitly
These samples should work in all versions, they're almost like
specifications (by example) and we should probably automate some
integration testing for each of them (thus ensuring it *actually* works
on all versions before releasing them).

In practice we don't have the time yet, so hopefully human testing will
catch any errors.

The argument goes this way: it is worse to have a silent bug in the
latest versions for a potentially long time and more likely to pop out
in production somewhere than it is to have a bug in the first samples a
user goes through. It doesn't make for a good initial experience but
that pressure to automate is good (and again better than the
alternative).