serve-folders: Move served dirs to dedicated subdir
Because the Docker builder Dockerfile syntax has no way to specify a
copy instruction for a directory (only for its contents) one needs to
effectively re-create a similarly named directory on the target and copy
the contents of the directory into it by specifying the same name in the
destination.

This makes it impossible to use multi-source COPY instructions.

The clean solution is to setup the build context accordingly beforehand
so that the contents of an arbitrary parent directory contain the
relevant directories and copy them all at once.