Enable plugins explicitly
17.10 enabled a breaking change whereby instead of loading as many
plugins as possible and introducing often convoluted ways of specifying
which plugin to use whenever it was conflicting with another or simply
not desired, now no plugins are loaded by default and all desired
plugins must be enabled explicitly with a simple *_ENABLED environment
variable or by providing an appropriate configuration file with a
conventional name (corresponding to that plugin).

Example of plugin using convention-based configuration file name:
serve-folders sample.  The samples modified here still use a single
orthanc.json and so are enabled using the environment variable.